LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/css-module/load-error-events-expected.txt

CONSOLE MESSAGE: TypeError: Import assertion type "css" is not valid
CONSOLE MESSAGE: TypeError: Import assertion type "css" is not valid
CONSOLE MESSAGE: TypeError: Import assertion type "css" is not valid
CONSOLE MESSAGE: TypeError: Import assertion type "css" is not valid
CONSOLE MESSAGE: TypeError: Import assertion type "css" is not valid
CONSOLE MESSAGE: TypeError: Import assertion type "css" is not valid
CONSOLE MESSAGE: TypeError: Import assertion type "css" is not valid
CONSOLE MESSAGE: TypeError: Import assertion type "css" is not valid

Harness Error (FAIL), message = TypeError: Import assertion type "css" is not valid

NOTRUN inline, 200, parser-inserted
NOTRUN inline, 404, parser-inserted
NOTRUN src, 200, parser-inserted
NOTRUN src, 404, parser-inserted
NOTRUN src, 200, not parser-inserted
NOTRUN src, 404, not parser-inserted
NOTRUN inline, 200, not parser-inserted
NOTRUN inline, 404, not parser-inserted
